fix htonl

+++ source/dnode/snode/src/snode.c
@@ -185,7 +185,7 @@ static int32_t buildStreamFetchRsp(SSDataBlock* pBlock, void** data, size_t* siz
   } else {
     pRetrieve->numOfRows = htobe64((int64_t)pBlock->info.rows);
     pRetrieve->numOfBlocks = htonl(1);
-    pRetrieve->numOfCols = htonl(taosArrayGetSize(pBlock->pDataBlock));
+    pRetrieve->numOfCols = htonl((uint32_t)taosArrayGetSize(pBlock->pDataBlock));
 
     int32_t actualLen = blockEncodeInternal(pBlock, pRetrieve->data + INT_BYTES * 2, blockSize, taosArrayGetSize(pBlock->pDataBlock));
     if (actualLen < 0) {
